feat: 당첨 번호는 1이상 45이하의 6개의 번호를 중복되지 않게 가지도록 한다.
- 기존의 숫자 검증 로직을 모두 NumberValidator 클래스로 이동
