+++ README.md
@@ -87,6 +87,7 @@
 - 당첨 번호를 확인한다.
     - [X] ~~지난 주 당첨 번호를 입력한다.~~
         - 쉼표 + 공백 (", ")를 통해서 번호를 구분한다.
+    - [X] ~~당첨 번호는 1이상 45이하의 6개의 번호를 중복되지 않게 가지도록 한다.~~
 - 당첨 통계를 구한다.
     - [ ] 발급한 로또가 당첨 번호와 몇개 일치 했는지 구한다.
     - [ ] 당첨 통계를 화면에 출력한다.
